ZNTA-1865 Extract Display project search result logic into new component file
